change data repository to fit new database schema

+++ Core/PeopleCountingStrategy.py
@@ -3,4 +3,4 @@ from Core import Image
 
 class PeopleCountingStrategy:
     @abstractmethod
-    def CountPeople(self, image: Image) -> int: pass
+    def CountPeople(self, image: Image) -> float: pass
